+++ src/lib/Tables.svelte
@@ -24,18 +24,17 @@
   <div class="hide-print">
     <h3>Show/Hide Question</h3>
     <div>
-      <input type="text" id="SelectQuestion" bind:value={text} />
-      <label for="SelectQuestion">ex : 12,13,15</label>
+      <input type="text" id="SelectQuestion" placeholder="ex : 12,13,15" bind:value={text} />
     </div>
     <ul>
       <li>
-        <input type="checkbox" bind:checked on:change={changeSelection} />
-        <span>All</span>
+        <input type="checkbox" id="show-all-q" bind:checked on:change={changeSelection} />
+        <label for="show-all-q">All</label>
       </li>
-      {#each questions as question (question.title)}
+      {#each questions as question,i (question.title)}
         <li>
-          <input type="checkbox" bind:checked={question.show} />
-          <span>{question.title}</span>
+          <input type="checkbox" id="{question.title}-{i}" bind:checked={question.show} />
+          <label for="{question.title}-{i}">{question.title}</label>
         </li>
       {/each}
     </ul>
@@ -43,17 +42,17 @@
 {/if}
 
 <style>
-  label {
-    margin-left: 4px;
-    filter: opacity(0.5);
+  input[type='text'] {
+    width: 99%;
   }
   h3 {
     margin: 0;
+    font-size: 18px;
   }
   ul {
     list-style: none;
-    padding-left: 0;
+    padding-left: 2px;
     overflow-y: scroll;
-    max-height: 45vh;
+    max-height: 40vh;
   }
 </style>